Move ProjectionBufferChanged.
