+++ src/InteractiveWindow/Editor/InteractiveWindow.cs
@@ -5,8 +5,6 @@
 
 using System;
 using System.Collections.Generic;
-using System.Collections.ObjectModel;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Threading;
@@ -35,9 +33,6 @@ namespace Microsoft.VisualStudio.InteractiveWindow
         private readonly IWpfTextView _textView;
 
         public event EventHandler<SubmissionBufferAddedEventArgs> SubmissionBufferAdded;
-
-        // the language engine and content type of the active submission:
-        private readonly IInteractiveEvaluator _evaluator;
 
         public PropertyCollection Properties { get; }
 
@@ -105,7 +100,7 @@ namespace Microsoft.VisualStudio.InteractiveWindow
 
         TextWriter IInteractiveWindow.ErrorOutputWriter => _errorOutputWriter;
 
-        IInteractiveEvaluator IInteractiveWindow.Evaluator => _evaluator;
+        IInteractiveEvaluator IInteractiveWindow.Evaluator => _dangerous_uiOnly.Evaluator; // Caller's responsibility
 
         /// <remarks>
         /// Normally, an async method would have an NFW exception filter.  This
@@ -431,383 +426,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow
 #endregion
 
 #region Buffers, Spans and Prompts
-        private object CreateStandardInputPrompt()
-        {
-            return string.Empty;
-        }
-
-        private object CreatePrimaryPrompt()
-        {
-            return _evaluator.GetPrompt();
-        }
-
-        private object CreateSecondaryPrompt()
-        {
-            // TODO (crwilcox) format prompt used to get a blank here but now gets "> " from get prompt.
-            return _evaluator.GetPrompt();
-        }
-
-        private ReplSpanKind GetSpanKind(SnapshotSpan span)
-        {
-            var textBuffer = span.Snapshot.TextBuffer;
-            if (textBuffer == _outputBuffer)
-            {
-                return ReplSpanKind.Output;
-            }
-            if (textBuffer == _standardInputBuffer)
-            {
-                return ReplSpanKind.StandardInput;
-            }
-            if (textBuffer.ContentType == _inertType)
-            {
-                return (span.Length == _lineBreakString.Length) && string.Equals(span.GetText(), _lineBreakString) ?
-                    ReplSpanKind.LineBreak :
-                    ReplSpanKind.Prompt;
-            }
-            return ReplSpanKind.Language;
-        }
-
-        /// <summary>
-        /// Return the index of the span containing the point. Returns the
-        /// length of the collection if the point is at the end of the last span.
-        /// </summary>
-        private int GetSourceSpanIndex(ReadOnlyCollection<SnapshotSpan> sourceSpans, SnapshotPoint point)
-        {
-            int low = 0;
-            int high = sourceSpans.Count;
-            while (low < high)
-            {
-                int mid = low + (high - low) / 2;
-                int value = CompareToSpan(_textView, sourceSpans, mid, point);
-                if (value == 0)
-                {
-                    return mid;
-                }
-                else if (value < 0)
-                {
-                    high = mid - 1;
-                }
-                else
-                {
-                    low = mid + 1;
-                }
-            }
-            Debug.Assert(low >= 0);
-            Debug.Assert(low <= sourceSpans.Count);
-            return low;
-        }
-
-        /// <summary>
-        /// Returns negative value if the point is less than the span start,
-        /// positive if greater than or equal to the span end, and 0 otherwise.
-        /// </summary>
-        private static int CompareToSpan(ITextView textView, ReadOnlyCollection<SnapshotSpan> sourceSpans, int index, SnapshotPoint point)
-        {
-            // If this span is zero-width and there are multiple projections of the
-            // containing snapshot in the projection buffer, MapUpToBuffer will return
-            // multiple (ambiguous) projection spans. To avoid that, we compare the
-            // point to the end point of the nearest non-zero width span instead.
-            int indexToCompare = index;
-            while (sourceSpans[indexToCompare].IsEmpty)
-            {
-                if (indexToCompare == 0)
-                {
-                    // Empty span at start of buffer. Point
-                    // must be to the right of span.
-                    return 1;
-                }
-                indexToCompare--;
-            }
-
-            var sourceSpan = sourceSpans[indexToCompare];
-            Debug.Assert(sourceSpan.Length > 0);
-
-            var mappedSpans = textView.BufferGraph.MapUpToBuffer(sourceSpan, SpanTrackingMode.EdgeInclusive, textView.TextBuffer);
-            Debug.Assert(mappedSpans.Count == 1);
-
-            var mappedSpan = mappedSpans[0];
-            Debug.Assert(mappedSpan.Length == sourceSpan.Length);
-
-            if (indexToCompare < index)
-            {
-                var result = point.CompareTo(mappedSpan.End);
-                return (result == 0) ? 1 : result;
-            }
-            else
-            {
-                var result = point.CompareTo(mappedSpan.Start);
-                if (result <= 0)
-                {
-                    return result;
-                }
-                result = point.CompareTo(mappedSpan.End);
-                return (result < 0) ? 0 : 1;
-            }
-        }
-
-        private const int SpansPerLineOfInput = 2;
-
-        private static readonly object s_suppressPromptInjectionTag = new object();
-
-        private bool TryGetCurrentLanguageBufferExtent(IProjectionSnapshot projectionSnapshot, out Span result)
-        {
-            if (projectionSnapshot.SpanCount == 0)
-            {
-                result = default(Span);
-                return false;
-            }
-
-            // the last source snapshot is always a projection of a language buffer:
-            var snapshot = projectionSnapshot.GetSourceSpan(projectionSnapshot.SpanCount - 1).Snapshot;
-            if (snapshot.TextBuffer != _currentLanguageBuffer)
-            {
-                result = default(Span);
-                return false;
-            }
-
-            SnapshotPoint start = new SnapshotPoint(snapshot, 0);
-            SnapshotPoint end = new SnapshotPoint(snapshot, snapshot.Length);
-
-            // projection of the previous version of current language buffer snapshot:
-            var surfaceSpans = projectionSnapshot.MapFromSourceSnapshot(new SnapshotSpan(start, end));
-
-            // the language buffer might be projected to multiple surface lines:
-            Debug.Assert(surfaceSpans.Count > 0);
-            result = new Span(surfaceSpans[0].Start, surfaceSpans.Last().End);
-            return true;
-        }
 
         private void ProjectionBufferChanged(object sender, TextContentChangedEventArgs e)
         {
-            // this is an edit performed in this event:
-            if (e.EditTag == s_suppressPromptInjectionTag)
-            {
-                return;
-            }
-
-            // projection buffer is changed before language buffer is created (for example, output might be printed out during initialization):
-            if (_currentLanguageBuffer == null)
-            {
-                return;
-            }
-
-            var oldSnapshot = (IProjectionSnapshot)e.Before;
-            var newSnapshot = (IProjectionSnapshot)e.After;
-            Span oldSurfaceLanguageBufferExtent;
-            Span newSurfaceLanguageBufferExtent;
-
-            if (!TryGetCurrentLanguageBufferExtent(oldSnapshot, out oldSurfaceLanguageBufferExtent) ||
-                !TryGetCurrentLanguageBufferExtent(newSnapshot, out newSurfaceLanguageBufferExtent))
-            {
-                return;
-            }
-
-            List<SpanRangeEdit> spanEdits = null;
-            var oldProjectionSpans = oldSnapshot.GetSourceSpans();
-            int oldProjectionSpanCount = oldProjectionSpans.Count;
-
-            // changes are sorted by position
-            foreach (var change in e.Changes)
-            {
-                // Old/new span might be outside of the language buffer -- on the left of it, since
-                // the language buffer always reaches to the end of the projection buffer.
-                Span oldSurfaceIntersection = oldSurfaceLanguageBufferExtent.Intersection(change.OldSpan) ?? new Span(oldSurfaceLanguageBufferExtent.Start, 0);
-                Span newSurfaceIntersection = newSurfaceLanguageBufferExtent.Intersection(change.NewSpan) ?? new Span(newSurfaceLanguageBufferExtent.Start, 0);
-
-                // change doesn't affect the language buffer:
-                if (oldSurfaceIntersection.IsEmpty && newSurfaceIntersection.IsEmpty)
-                {
-                    continue;
-                }
-
-                var oldSurfaceStartLineNumber = oldSnapshot.GetLineNumberFromPosition(oldSurfaceIntersection.Start);
-                var oldSurfaceEndLineNumber = oldSnapshot.GetLineNumberFromPosition(oldSurfaceIntersection.End);
-
-                // The change doesn't include line breaks and is entirely within the current
-                // language buffer. Note that we might need to proceed with span replacement even if
-                // line count delta is zero: the tracking spans of all but last language buffer
-                // projections need to stop growing.
-                if (oldSurfaceStartLineNumber == oldSurfaceEndLineNumber &&
-                    change.LineCountDelta == 0 &&
-                    oldSurfaceIntersection == change.OldSpan &&
-                    newSurfaceIntersection == change.NewSpan)
-                {
-                    continue;
-                }
-
-                // calculate which prompts and language projection spans to remove and replace with new spans:
-                int oldStartSpanIndex = GetProjectionSpanIndexFromEditableBufferPosition(oldSnapshot, oldProjectionSpanCount, oldSurfaceStartLineNumber);
-                int oldEndSpanIndex = GetProjectionSpanIndexFromEditableBufferPosition(oldSnapshot, oldProjectionSpanCount, oldSurfaceEndLineNumber);
-
-                int spansToReplace = oldEndSpanIndex - oldStartSpanIndex + 1;
-                Debug.Assert(spansToReplace >= 1);
-
-                var newSubjectStartLine = newSnapshot.MapToSourceSnapshot(newSurfaceIntersection.Start).GetContainingLine();
-                var newSubjectEndLine = newSnapshot.MapToSourceSnapshot(newSurfaceIntersection.End).GetContainingLine();
-
-                var newSubjectEndLineNumber = newSubjectEndLine.LineNumber;
-
-                int i = 0;
-                int lineBreakCount = newSubjectEndLineNumber - newSubjectStartLine.LineNumber;
-                var newSpans = new object[lineBreakCount * SpansPerLineOfInput + 1];
-
-                var subjectLine = newSubjectStartLine;
-                while (true)
-                {
-                    if (subjectLine.LineNumber != newSubjectStartLine.LineNumber)
-                    {
-                        // TODO (crwilcox): do we need two prompts?  Can I tell it to not do this?  Or perhaps we do want this since we want different markings?
-                        newSpans[i++] = CreateSecondaryPrompt();
-                    }
-
-                    newSpans[i++] = CreateLanguageSpanForLine(subjectLine);
-
-                    if (subjectLine.LineNumber == newSubjectEndLineNumber)
-                    {
-                        break;
-                    }
-
-                    subjectLine = subjectLine.Snapshot.GetLineFromLineNumber(subjectLine.LineNumber + 1);
-                }
-
-                Debug.Assert(i == newSpans.Length);
-
-                if (spanEdits == null)
-                {
-                    spanEdits = new List<SpanRangeEdit>();
-                }
-
-                spanEdits.Add(new SpanRangeEdit(oldStartSpanIndex, spansToReplace, newSpans));
-            }
-
-            if (spanEdits != null)
-            {
-                ReplaceProjectionSpans(oldProjectionSpans, spanEdits);
-            }
-
-            CheckProjectionSpans();
-        }
-
-        /// <remarks>
-        /// This should only be called from within the current language buffer.  If there are
-        /// any output or standard input buffers between the specified line and the end of the
-        /// surface buffer, then the result will be incorrect.
-        /// </remarks>
-        private int GetProjectionSpanIndexFromEditableBufferPosition(IProjectionSnapshot surfaceSnapshot, int projectionSpansCount, int surfaceLineNumber)
-        {
-            // The current language buffer is projected to a set of projections interleaved regularly by prompt projections 
-            // and ending at the end of the projection buffer, each language buffer projection is on a separate line:
-            //   [prompt)[language)...[prompt)[language)<end of projection buffer>
-            int result = projectionSpansCount - (surfaceSnapshot.LineCount - surfaceLineNumber) * SpansPerLineOfInput + 1;
-            Debug.Assert(GetSpanKind(surfaceSnapshot.GetSourceSpan(result)) == ReplSpanKind.Language);
-            return result;
-        }
-
-        private void ReplaceProjectionSpans(ReadOnlyCollection<SnapshotSpan> oldProjectionSpans, List<SpanRangeEdit> spanEdits)
-        {
-            Debug.Assert(spanEdits.Count > 0);
-
-            int start = spanEdits[0].Start;
-            int end = spanEdits[spanEdits.Count - 1].End;
-
-            var replacement = new List<object>();
-            replacement.AddRange(spanEdits[0].Replacement);
-            int lastEnd = spanEdits[0].End;
-
-            for (int i = 1; i < spanEdits.Count; i++)
-            {
-                SpanRangeEdit edit = spanEdits[i];
-
-                int gap = edit.Start - lastEnd;
-
-                // there is always at least prompt span in between subsequent edits
-                Debug.Assert(gap != 0);
-
-                // spans can't share more then one span
-                Debug.Assert(gap >= -1);
-
-                if (gap == -1)
-                {
-                    replacement.AddRange(edit.Replacement.Skip(1));
-                }
-                else
-                {
-                    replacement.AddRange(oldProjectionSpans.Skip(lastEnd).Take(gap).Select(CreateTrackingSpan));
-                    replacement.AddRange(edit.Replacement);
-                }
-
-                lastEnd = edit.End;
-            }
-
-            _projectionBuffer.ReplaceSpans(start, end - start, replacement, EditOptions.None, s_suppressPromptInjectionTag);
-        }
-
-        private object CreateTrackingSpan(SnapshotSpan snapshotSpan)
-        {
-            var snapshot = snapshotSpan.Snapshot;
-            if (snapshot.ContentType == _inertType)
-            {
-                return snapshotSpan.GetText();
-            }
-            return new CustomTrackingSpan(snapshot, snapshotSpan.Span, PointTrackingMode.Negative, PointTrackingMode.Negative);
-        }
-
-        private ITrackingSpan CreateLanguageSpanForLine(ITextSnapshotLine languageLine)
-        {
-            var span = languageLine.ExtentIncludingLineBreak;
-            bool lastLine = (languageLine.LineNumber == languageLine.Snapshot.LineCount - 1);
-            return new CustomTrackingSpan(
-                _currentLanguageBuffer.CurrentSnapshot,
-                span,
-                PointTrackingMode.Negative,
-                lastLine ? PointTrackingMode.Positive : PointTrackingMode.Negative);
-        }
-
-        // Verify spans and GetSourceSpanIndex.
-        [Conditional("DEBUG")]
-        private void CheckProjectionSpans()
-        {
-            var snapshot = _projectionBuffer.CurrentSnapshot;
-            var sourceSpans = snapshot.GetSourceSpans();
-            int n = sourceSpans.Count;
-
-            // Spans should be contiguous and span the entire buffer.
-            int offset = 0;
-            for (int i = 0; i < n; i++)
-            {
-                // Determine the index of the first non-zero width
-                // span starting at the same point as current span.
-                int expectedIndex = i;
-                while (sourceSpans[expectedIndex].IsEmpty)
-                {
-                    expectedIndex++;
-                    if (expectedIndex == n)
-                    {
-                        break;
-                    }
-                }
-                // Verify GetSourceSpanIndex returns the expected
-                // index for the start of the span.
-                int index = GetSourceSpanIndex(sourceSpans, new SnapshotPoint(snapshot, offset));
-                Debug.Assert(index == expectedIndex);
-                // If this is a non-empty span, verify GetSourceSpanIndex
-                // returns the index for the midpoint of the span.
-                int length = sourceSpans[i].Length;
-                if (length > 0)
-                {
-                    index = GetSourceSpanIndex(sourceSpans, new SnapshotPoint(snapshot, offset + length / 2));
-                    Debug.Assert(index == i);
-                }
-                offset += length;
-            }
-
-            Debug.Assert(offset == snapshot.Length);
-
-            if (n > 0)
-            {
-                int index = GetSourceSpanIndex(sourceSpans, new SnapshotPoint(snapshot, snapshot.Length));
-                Debug.Assert(index == n);
-            }
+            UIThread(uiOnly => uiOnly.ProjectionBufferChangedInternal(sender, e));
         }
 
         #endregion
